Switched uses of 'os.path.split' in order to get filenames to 'os.path.basename'.
Created an OutputFileExistsDialog class giving the user an option to replace or rename a new output file if one with the chosen name already exists.

Added a feature that resizes the font on buttons.

Added an option to configure where the program saves output files and remember this between sessions.